requirements.txt
jinja2==3.1.2
markdown==3.5.1
pyyaml==6.0.1
watchdog==6.0.0
cairosvg==2.8.2
Pillow==12.0.0
